# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Live Test",
        "description": "This is api testing",
        "contact": {
            "email": "xyz@xyz.com"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/friend/referral/create": {
            "post": {
                "tags": [
                    "Friend"
                ],
                "summary": "create referral friend",
                "operationId": "create-referral-friend",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/ticket/create": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "create ticket",
                "operationId": "create-ticket",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/ticket/send": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Send ticket to customers",
                "operationId": "send-ticket-to-customer",
                "parameters": [
                    {
                        "name": "mytest",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}